chore(sweep): dead code out, docs filed, flutter_webrtc trimmed to Android, dart fix applied
Pre-Play sweep of the whole tree (BRAIN §251).

- ~30 dead Dart members and core/realtime/screen_presence.dart removed, each
  verified unreferenced in lib/ and test/ by a tokenizing scan and re-scanned
  for cascades.
- `dart fix --apply`: 249 fixes in 117 files; analyzer infos 599 -> 201. Two
  regressions from omit_local_variable_types fixed and commented
  (encrypted_media_cache._provider, call_controller._ensureRelay).
- third_party/flutter_webrtc: ios/, macos/, linux/, windows/, common/,
  third_party/, docs, sponsor assets and tooling removed (340 -> 105 files);
  plugin pubspec platforms: android only. android/ untouched.
- docs/: architecture set and fifteen superseded guides moved to
  docs/archive/; stale instructions copy, both audit JSONs (one carried the
  burned Maps key) and docs/legal/ (drifted copies of web/) deleted.
  REFERENCE.md rewritten from the tree; root, mobile and web READMEs
  rewritten; every dangling docs/ path repointed.
- Dead Google Maps config removed (.env.example key, maps.properties example
  and ignore rule); .cursor/ untracked and ignored; launch.json trimmed.

Gates at this tree: flutter analyze 0 errors / 0 warnings (201 infos);
flutter test 1464 passed inside `tool/release.sh --play`, which then built
the play AAB for build 73 (sha256 802e19b2...) with all three libapp.so
stamped miles-build-73; `:flutter_webrtc:compileDebugJavaWithJavac` green.

Left for others: nothing pending from the concurrent session (its §250 work
landed in 3238b5f; the lint hunks here sit on top of it). The build-73 play
APK it left under build/ was preserved outside the tree before the clean.

+++ mobile/lib/features/unlink/scene/unlink_end_overlay.dart
@@ -157,8 +157,8 @@ class _UnlinkEndOverlayState extends State<UnlinkEndOverlay>
                 fit: BoxFit.cover,
                 clipBehavior: Clip.hardEdge,
                 child: SizedBox(
-                  width: film.value.size.width.toDouble(),
-                  height: film.value.size.height.toDouble(),
+                  width: film.value.size.width,
+                  height: film.value.size.height,
                   child: VideoPlayer(film),
                 ),
               );
